pyproject.toml
[tool.poetry]
name = "django-simulation-project"
version = "0.1.0"
description = "A Django project for simulation and parameter processing"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
djangorestframework = "^3.14"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
celery = "^5.2.7"
redis = "^4.5.1"
django-cors-headers = "^4.0.0"
drf-spectacular = "^0.27.0"
XlsxWriter = "^3.1.0"
python-magic = "^0.4.27"
numpy = "^1.24.0"
pandas = "^2.0.0"
openpyxl = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-django = "^4.5.2"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v -p no:warnings --strict-markers"
testpaths = ["tests"]
